file_util: Throw exception, if file can't be read
Before file_util::contents would just return an empty string if the file
couldn't be opened, because std::getline would evaluate to false right
away.

Every caller of file_util::contents now needs to potentially handle a
thrown exception. Right now the only callers are modules, the handler
for `include-file` directive and the handler for the ${file:...}
reference (dereference_file).

In general, if a module fails to read a file it needs to read, it can
let the exception bubble up the call stack and let the controller stop
the module. If a read error is recoverable, the module has to catch the
exception itself; right now there aren't any modules where this would be
recoverable.

If an include-file cannot be read, polybar will terminate, which is the
same behavior as when the include-file doesn't exist

As for dereference_file, currently, if any read error occurs, the
application error is always thrown. It may be desirable to implement the
same behavior as when the file doesn't exist (use a fallback).