feat(network): toggleable labels
Added new labels for altering displayed information in connected and
packetloss states. Labels are toggled by clicking on the battery module
in any place. It is not possible to click if after altering information
may become empty. Formatter has new check for static (non-tag) content. It
is used to be sure that labels can be safely toggled.