refactor(tray): Allow turning off reparenting
Introduces `tray-reparent = true` in the bar section

Reparenting may impede our ability to use the tray offset values to
their fullest extent. For example in dropbox, the bar parent window is a
window with the exact same dimensions, so if the tray becomes a child of
that window, we cannot get the tray to draw outside of those bounds as
we could before.