i3: Add possibility to sort workspaces by name
In i3 it is possible to assign workspaces with arbitrary names without
assigning it a specific index. Polybar however, is only able to sort the
workspaces by index, leaving named workspaces in seemingly random order
received from i3.

Make it possible to configure sorting on workspace names by adding an
option ("name-sort") for the i3 module.