fix(file): Don't add slash to relative path
This fixes a regression introduced in
56e24992dfe2fba32b1f6b07fc37ab7d0a7c6aba where relative config file
paths weren't recognized because file_util::expand just added a slash
to the beginning.

That is calling `polybar -c config example` would try to load the config
file at `/config` instead of using the relative path as before.

In all other cases where expand is used this change shouldn't matter
because polybar only accepts absolute paths everyhwere else.
Theoretically this would now allow relative paths (relative to the cwd
where polybar was called) but this shouldn't used (or documented)
because that behavior will change when merging #1523 which would make
paths relative to the polybar config.

Ref #1523
Ref 56e24992dfe2fba32b1f6b07fc37ab7d0a7c6aba