feat: Add option to not enforce unique module names
Adds a new config key `allow-multiple` to every module section, by
default it's set to `false`.

If set to true, the controller will not emit an error when that module
appears multiple times.

There are still some cases where using the same module multiple times is
desirable. For example when you want to have different separators on the
left side than on the right side. adi1090x's polybar-themes does this
in several places.

Since in the future the module name is used to target actions, this will
lead to unspecified behavior when such a module receives an action.
I think when sending actions to modules, we will check allow_multiple()
and either not deliver the action at all or just emit an error.

Ref: https://github.com/adi1090x/polybar-themes