Right-click on pulseaudio launches control panel
Adds a new EVENT_OPEN_CONTROL_PANEL event to the pulseaudio module.

The event is bound to right-click, and it's handler invokes the
"control-panel" app as defined in the config file (default is
"pavucontrol")

If the specified control panel command does not end in "&", one is
appended so that the system() all will not block.