Implementing a flag that enables setting dealing with absent interfaces as being functionally present.
This is useful if ephemeral interfaces exist, ie. if interfaces are not guaranteed to be present
throughout the entire time polybar is being run. Setting the flag to true enables polybar to
continue polling absent interfaces.

It's possible to add a `ignore-not-present` boolean field to the polybar config file under the
specific network interface to enable this feature.