Add basic version of hover actions
Added HOVER to the mousebtn enum and the logic in bar.cpp to handle hover actions.
In the mouse motion event handler, there's now a check for actions with the HOVER mask.
Bar now keeps track of the last action string detected in the hover check. If the action detected is different from the las action, both actions are called. Otherwise, nothing happens. This means that hover works like a toggle.
I'm currently bipassing the check for whether to listen to motion events or not with a '|| true'. This will be replaced by a better alternative later on.
Also added hover action support to custom/script.